+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Skip conversion for element name only message targets
 - Replace deprecated `substr` usages
 - Check ModelProviderService before access
+- Service annotation `@cov2ap.ignore` to exclude service from OData V2 adapter
+- Log error with statusCode 400 as warning (before it was error)
 
 ## Version 1.11.9 - 2023-11-20